@@ -1,16 +1,18 @@
 ﻿using System;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
+
 namespace Extensions
 {
-    public static class IWebHostBuilderExtensions
+    public static class IHostBuilderExtensions
     {
-        public static IWebHostBuilder UseConfigFile(this IWebHostBuilder builder, string fileName, bool optional = false)
+        public static IHostBuilder UseConfigFile(this IHostBuilder builder, string fileName, bool optional = false)
         {
             try
             {
                 return builder
-                    .ConfigureAppConfiguration(config => config.AddJsonFile(fileName, optional));
+                    .ConfigureAppConfiguration((ctx, config) => config.AddJsonFile(fileName, optional));
             }
             catch (Exception ex)
             {
